# enhance: mini optimization in looking up notes by filename