# enhance: mini optimization in looking up notes by filename
## General

### Quality Assurance
- [x] make sure all the existing [tests](https://wiki.dendron.so/notes/cb22bd36-d45a-4ffd-a31e-96c4b39cb7fb.html#running-all-tests) pass
- [x] do a spot check by running your feature with our [test workspace](https://wiki.dendron.so/notes/cb22bd36-d45a-4ffd-a31e-96c4b39cb7fb.html#test-workspace)
- [] after you submit your pull request, check the output of our [integration test](https://github.com/dendronhq/dendron/actions) and make sure all tests pass
  - NOTE: if you running mac/linux, check the windows output and vice versa if you are developing on windows

### Details
Small optimization, however, this method is used in many places. 

Some benchmarking shows almost ~50%+ speed up of this function.

```
Sample id being compared: '0d342ce5-3648-472a-85eb-5a447de1be091fb8b41c-d7f6-4a46-870c-dc6b1a92023f7904d17c-66dc-413e-9f53-bed41cb1cedc'

Took '7ms' to loop through and equal compare '100000' when stacked '3' in single id, (toLowerInEachIteration: 'false)'
Took '12ms' to loop through and equal compare '100000' when stacked '3' in single id, (toLowerInEachIteration: 'true)'
```
Running benchmark while computer was busy:
```
Sample id being compared: '376abe8e-0c0e-4f03-9a52-9050a4ca5b962469b75b-ee93-4923-9ae1-783cd1d69e0f37d08237-a903-4e0f-a090-6d39a8497afd'
Took '26ms' to loop through and equal compare '100000' when stacked '3' in single id, (toLowerInEachIteration: 'false)'
Took '50ms' to loop through and equal compare '100000' when stacked '3' in single id, (toLowerInEachIteration: 'true)'
```

Benchmark code:
```
import { v4 as uuidv4 } from 'uuid';

export const uuid = () => {
  return uuidv4();
};

function createSingleId(howManyUuidsInOneId1: number) {
  let id = '';
  for (let j = 0; j < howManyUuidsInOneId1; j++) {
    id += uuid();
  }
  return id;
}

function generate(howManyIds: number, howManyUuidsInOneId: number) {
  const ids = [];
  for (let i = 0; i < howManyIds; i++) {
    ids.push(createSingleId(howManyUuidsInOneId));
  }
  return ids;
}

const main = (howManyIds: number, howManyUuidsInOneId: number, toLowerInEachIteration: boolean) => {

  const ids = generate(howManyIds, howManyUuidsInOneId);

  const toFind = createSingleId(howManyUuidsInOneId);

  console.log(`Sample id being compared: '${toFind}'`);

  const before = new Date().getTime();

  if (toLowerInEachIteration) {
    ids.filter(id => id === toFind.toLowerCase());
  } else {
    ids.filter(id => id === toFind);
  }

  console.log(`Took '${new Date().getTime() - before}ms' to loop through and equal compare '${howManyIds}' when stacked '${howManyUuidsInOneId}' in single id, (toLowerInEachIteration: '${toLowerInEachIteration})'`);
};

main(100000, 3, false);
main(100000, 3, true);
```